Moved the .ino file to its own directory
The Arduino app insists on the ino file being in a folder with the same base name.

This way, the entire tree can be zipped up and installed using Sketch>Include Library>Add ZIP Library

Also changed the #include directive to what the Arduino tool uses when you use "Include library". Not sure if that makes a difference.