available( ) should check for RSR[RXEMPTY] instead of SR[WCF]
The Word Complete Flag goes high only when an entire word has been received and probably goes low whenever another bit comes in for the next word.

The SR[RXD] could also be used to check if data is available, but only if the high-water setting for the receive FIFO is set to 0.