Replaced macros with struct fields; fixed bugs
Removed the base class with the ISR function. I suspect this was an attempt to use the same code for all interrupt service handlers, but there's just no easy way around the fact that the interrupt handler can't have a "this". So now there is an interrupt service routine for each port, and a switch-case in the constructor to set things up.

The constructor now takes a port number as parameter, which is used to select a case in a switch-case statement. I may change that into a const struct or class (like the TeensyDuino SPI library does). For now, only portnum=0 is supported. SPI1, SPI2 and SPI3 will be added later.

All the SPICLASS_xxx macros were replaced with statements that access the fields of the _lpspi struct pointer which is initialized at construction time. Also, all the "magic numbers" to put values in bit fields of _lpspi were replaced with macros from the IMXRT.h file. The only thing that's still hardcoded is the code to initialize the "daisy" and "mux" registers to set up the port. I'll either have to read the datasheet about those, or import some code from the TeensyDuino SPI library.

BUGFIX: Changed the internal data dump function: It was conditional upon FCF (Frame Complete, i.e. PCS asserted) and WCF (Word Complete, i.e. an entire word is in the incoming shift register), which didn't make sense at all: First of all, unless I misunderstand the data sheet, WCF is only true when the last bit of the incoming data is in the shift register, so when the next word starts coming in (and the FIFO is perfectly valid), WCF is false and the function didn't process the incoming data.

Secondly: if the function only handles incoming data when FCF is asserted, the incoming message must fit into the RX FIFO, otherwise it will overrun the SPI port.

Another problem was that the code would echo the incoming data to the output (SLAVE_TDR = val), and this corrupted the incoming data. I don't know why.

Clearing the WCF flag also doesn't make sense: there's no extensive documentation but I suspect that doing this while a word is coming in, will discard some bits of the word because the shift register is probably reset by clearing WCF.

So the internal data dump function now prints as many incoming words as it can, regardless of whether the CS line is active or not. The current code sets the high-water mark for the receive interrupt to 0, so the interrupt service routine is called for each incoming word. But if you change the high-water mark to a different value, the handler gets called less often, and has to process more input at a time. This could be beneficial for some applications. Keep in mind however that if the high-water mark is nonzro and there's data in the FIFO but no more data comes in from the master, the IRQ is never called. To fix that, you would have to also program a timer that processes the FIFO if no data comes in for a while.

BUGFIX: Previously, the interrupt was triggered on a transmit (and the initialization generates the first transmit by sending a 0, and the receive interrupt handlers would cause a transmit to happen too). This caused a problem in my test setup: The last byte of the incoming message would not show up until the beginning of the next message. I changed the IRQ initialization so that it uses the RDIE flag, not the TDIE flag of the IER register, which fixed that problem. I'm not sure if this is why the read-data gets corrupted if you send an outgoing word in the interrupt service routine, but it could be related.

I simplified the demo code (spi_slave.ino) a little bit, so the interrupt routine simply prints values in hexadecimal and doesn't skip the dump code if active() returns false. It also doesn't output the array to the output anymore.